Explicit boxing
git-svn-id: https://svn.apache.org/repos/asf/commons/proper/dbutils/branches/2_0@1457563 13f79535-47bb-0310-9956-ffa450edef68

+++ src/main/java/org/apache/commons/dbutils2/AbstractExecutor.java
@@ -183,7 +183,7 @@ abstract class AbstractExecutor<T extends AbstractExecutor<T>> {
 
         // go through and bind all of the positions for this name
         for (Integer p:pos) {
-            stmt.setNull(p, sqlType);
+            stmt.setNull(p.intValue(), sqlType);
         }
 
         // add the param and value to our map
@@ -216,7 +216,7 @@ abstract class AbstractExecutor<T extends AbstractExecutor<T>> {
         // go through and bind all of the positions for this name
         for (Integer p:pos) {
             // TODO: need to figure out how to bind NULL
-            stmt.setObject(p, value);
+            stmt.setObject(p.intValue(), value);
         }
 
         // add the param and value to our map
